(#237)refactor: 피드백 service interface 구현했습니다.
